доработала контроллер

+++ activity/src/main/java/activityApp/service/ActivityService.java
@@ -5,5 +5,5 @@ import activityApp.domain.Activity;
 import java.util.List;
 
 public interface ActivityService {
-    List<Activity> getActivitiesByTripId(Integer before, Integer after);
+    List<Activity> getActivitiesByTripId(long before, long after);
 }
